requirements.txt
Django==3.2.25
djangorestframework==3.15.1
django-extensions==3.0.0
psycopg2-binary==2.9.9
django-filter==2.3.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2023.6.1
